Refactor: move `"use cache"` revalidation logic out of incremental cache
The `revalidateTag` method of the incremental cache should not also
handle the revalidation for the `"use cache"` cache handlers.

This PR relocates those bits to the revalidation utils module, which
was previously used for `after`, and has now been lifted one level up.

There should be no behavioral change in this refactoring.
