+++ packages/next/src/server/after/after-context.ts
@@ -4,7 +4,7 @@ import type { AfterCallback, AfterTask } from './after'
 import { InvariantError } from '../../shared/lib/invariant-error'
 import { isThenable } from '../../shared/lib/is-thenable'
 import { workAsyncStorage } from '../app-render/work-async-storage.external'
-import { withExecuteRevalidates } from './revalidation-utils'
+import { withExecuteRevalidates } from '../revalidation-utils'
 import { bindSnapshot } from '../app-render/async-local-storage'
 import {
   workUnitAsyncStorage,
