+++ packages/next/src/server/app-render/action-handler.ts
@@ -49,6 +49,7 @@ import { synchronizeMutableCookies } from '../async-storage/request-store'
 import type { TemporaryReferenceSet } from 'react-server-dom-webpack/server.edge'
 import { workUnitAsyncStorage } from '../app-render/work-unit-async-storage.external'
 import { InvariantError } from '../../shared/lib/invariant-error'
+import { executeRevalidates } from '../revalidation-utils'
 
 function formDataFromSearchQueryString(query: string) {
   const searchParams = new URLSearchParams(query)
@@ -517,15 +518,7 @@ export async function handleAction({
   requestStore.phase = 'action'
 
   const resolvePendingRevalidations = async () =>
-    workUnitAsyncStorage.run(requestStore, () =>
-      Promise.all([
-        workStore.incrementalCache?.revalidateTag(
-          workStore.pendingRevalidatedTags || []
-        ),
-        ...Object.values(workStore.pendingRevalidates || {}),
-        ...(workStore.pendingRevalidateWrites || []),
-      ])
-    )
+    workUnitAsyncStorage.run(requestStore, () => executeRevalidates(workStore))
 
   // When running actions the default is no-store, you can still `cache: 'force-cache'`
   workStore.fetchCache = 'default-no-store'
